Revert "Feature/get position"
