--- requirements.txt
@@ -1,2 +1,0 @@
-requests>=2.31.0,<3
-python-dotenv>=1.0.1,<2
